+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [1.0.1] - 2026-01-18
+
+### Added
+* Creating upload-all script for conan packages.
+
+### Changed
+* Updating setup scripts to wire up a conan cache.
+
 ## [1.0.0] - 2026-01-11
 
 ### Changed